add code generator, some refactoring
