+++ .scalafix.conf
@@ -1,13 +1,9 @@
 rules = [
   DisableSyntax
-  ExplicitResultTypes
   LeakingImplicitClassVal
   NoAutoTupling
   NoValInForComprehension
   OrganizeImports
-  ProcedureSyntax
-  RemoveUnused
-  MissingFinal
 ]
 
 OrganizeImports {
@@ -21,10 +17,6 @@ OrganizeImports {
     "*",
     "zio."
   ]
-}
-
-RemoveUnused {
-  imports = false # handled by OrganizeImports
 }
 
 DisableSyntax.noThrows = true
