+++ .scalafix_test.conf
@@ -1,22 +1,10 @@
 rules = [
-  Disable
   DisableSyntax
-  ExplicitResultTypes
   LeakingImplicitClassVal
   NoAutoTupling
   NoValInForComprehension
   OrganizeImports
-  ProcedureSyntax
-  RemoveUnused
-  MissingFinal
 ]
-
-Disable {
-  ifSynthetic = [
-    "scala/Option.option2Iterable"
-    "scala/Predef.any2stringadd"
-  ]
-}
 
 OrganizeImports {
   removeUnused = true
@@ -31,10 +19,8 @@ OrganizeImports {
   ]
 }
 
-RemoveUnused {
-  imports = false # handled by OrganizeImports
-}
-
+DisableSyntax.noThrows = false
+DisableSyntax.noNulls = false
 DisableSyntax.noReturns = true
 DisableSyntax.noXml = true
 DisableSyntax.noFinalize = true
