+++ docker-compose.yaml
@@ -1,7 +1,7 @@
 services:
   pubsub:
     # https://console.cloud.google.com/gcr/images/google.com:cloudsdktool/GLOBAL/cloud-sdk
-    image: gcr.io/google.com/cloudsdktool/cloud-sdk:480.0.0-emulators
+    image: gcr.io/google.com/cloudsdktool/cloud-sdk:498.0.0-emulators
     ports:
       - "8085:8085"
     command: gcloud beta emulators pubsub start --project=any --host-port=0.0.0.0:8085
