+++ project/build.properties
@@ -1,1 +1,1 @@
-sbt.version = 1.10.3
+sbt.version = 1.10.4
