+++ zio-pubsub/shared/src/main/scala/com/anymindgroup/pubsub/model/PubsubConnectionConfig.scala
@@ -5,8 +5,8 @@ sealed trait PubsubConnectionConfig {
 }
 
 object PubsubConnectionConfig {
-  final case class Cloud(project: GcpProject)                  extends PubsubConnectionConfig
-  final case class Emulator(project: GcpProject, host: String) extends PubsubConnectionConfig
+  final case class Cloud(project: GcpProject)                                    extends PubsubConnectionConfig
+  final case class Emulator(project: GcpProject, host: String, port: Int = 8085) extends PubsubConnectionConfig
 
   final case class GcpProject(name: String) extends AnyVal {
     override def toString: String = name
