+++ zio-pubsub/shared/src/main/scala/com/anymindgroup/pubsub/model/TopicName.scala
@@ -1,0 +1,6 @@
+package com.anymindgroup.pubsub.model
+
+final case class TopicName(projectId: String, topic: String) {
+  def path: String              = s"projects/$projectId/topics/$topic"
+  override def toString: String = path
+}
